fix(domain): align double_spend_proof version parameter with sibling message classes (#283)
`double_spend_proof::to_data(size_t)` and `serialized_size(size_t)`
were the only two call sites in the message family that used
`size_t` for the protocol version. Every sibling (`merkle_block`,
`compact_block`, `get_blocks`, `get_headers`,
`prefilled_transaction`, `double_spend_proof::spender`, ...) takes
`uint32_t`, and so does `double_spend_proof::from_data` — the
version field itself is declared `uint32_t version = 0;` inside
`spender`.

The inconsistency is load-bearing downstream: the C-API generator
pipes `size_t` through as `kth_size_t` and `kth_size_t` maps to
`Py_ssize_t` on the py-native side, so the Python wrapper for DSP
ended up parsing a non-negative protocol version as a signed
Py_ssize_t with a non-negativity check that's structurally
unreachable for an unsigned type. Every other message binding
already parses it as `unsigned int` / `"I"` and casts to
`uint32_t`.

Change the four DSP entry points — `to_data(uint32_t) const`,
`to_data(uint32_t, data_sink&) const`, the templated
`to_data(uint32_t, W&)`, and `serialized_size(uint32_t) const` —
to take `uint32_t`, matching the sibling classes. Implementation
signatures in `double_spend_proof.cpp` follow the header.

Tests call `dsp.to_data(0)` / `dsp.serialized_size(0)` with the
integer literal `0`, which works unchanged under either type, so
the change is source-compatible for in-tree callers.

C-API regenerated through the generator pipeline: the three entry
points that previously carried `kth_size_t version` now carry
`uint32_t version`, matching the signature shape that
`merkle_block`, `compact_block`, `get_blocks`, `get_headers`,
`prefilled_transaction` and `double_spend_proof::spender`
already ship.

+++ src/domain/src/message/double_spend_proof.cpp
@@ -107,7 +107,7 @@ expect<double_spend_proof> double_spend_proof::from_data(byte_reader& reader, ui
 // Serialization.
 //-----------------------------------------------------------------------------
 
-data_chunk double_spend_proof::to_data(size_t version) const {
+data_chunk double_spend_proof::to_data(uint32_t version) const {
     data_chunk data;
     auto const size = serialized_size(version);
     data.reserve(size);
@@ -118,7 +118,7 @@ data_chunk double_spend_proof::to_data(size_t version) const {
     return data;
 }
 
-void double_spend_proof::to_data(size_t version, data_sink& stream) const {
+void double_spend_proof::to_data(uint32_t version, data_sink& stream) const {
     ostream_writer sink_w(stream);
     to_data(version, sink_w);
 }
